fixed the code to suit norminette rules and also the header file as well.
the Makefile now includes the -f flag when using rm, in order to silence the errors when there is no object file to be removed.

furthermore, some minor improvements were made to handle the case of 0 philosophers. Old version had an issue where the killer process would be made, and the main would wait for one of the philosophers to join.